Add function to map of expressions in a select query
This is a combination of lenses of map functions. It's quite unsatisfying actually.
The Lens library by Obrok in Elixir allowed us to make deeply nested changes all over
the place in a structure. Haskell and correspondingly F# lenses and optics seem a lot
more strict.

Lenses, and all the infrastructure around them, are expected to abide by the following
three laws:

1) You get back what you put in
2) Putting back what you got doesn't change anything
3) Setting twice is the same as setting once

This in turn makes any lens that is focusing on say all expressions within an
expression tree problematic. Consider the lens `allLeafExpressions` which given
an expression tree returns all the leaves. The return type of this lens when used
as a getter (Lens.get/Lens.view/Lens.preview) or equivalent returns a sequence
of leaf nodes, which is strictly different from the structure of the query tree,
and therefore cannot be put back into the data structure, hence violating law 1.

I believe these are called traversals or so, but I haven't been able to get them
to work with a "change everything in place" style lens like what I was after.

In this particular instance I wanted the ability to map every expression in a query.
An expression might be in a WHERE-clause, a select column, or in any one of a number
of other places. I ended up having to build this as a combination of lenses
and a custom mapping function.