Use junk to do low count checks
The analyzer inserts a junk producing Diffix Specific aggregate
into queries that do anonymization. For queries that do not
themselves already aggregate, it also marks the query as later
needing to be expanded again.

The junk aggregator spits out whether a set of attributes are
low count or not, and furthermore what the user count is. The
latter information is later used in the optional expansion plan
note to replicate rows.

The custom plan type might not transfer to Postgres, but it made
the implementation easier.