Changes for hosting under github pages
We are hosted at /website/playground.
This breaks the integrity checking done by
blazor on boot, since the hashes don't match
because they are path dependent.

For now I have disabled the integrity checking.
In the future we are likely going to be better
off serving this tool under the root directory
of a site, such as playground.open-diffix.org