Make count distinct distribution algorithm deterministic
Sorts input data by both the number of values, as well as the AID value itself to get a consistent ordering within a group of users with the same number of values.

Sort the values of an individual AID to make sure the order in which values are picked is deterministic as well.