Drop buckets where any one of the AIDs has insufficient data
For a while I thought it was OK for an AID to altogether have no data at all. This turned out to be a flawed assumption as this would lead to the difference between 1 and 0 AID values for an AID being clearly visible. The design has now reverted to the previous one whereby it is necessary for each AID to enough AID values to produce a sensible anonymous aggregate.

In the process I realized that the the order of the AID columns would impact the anonymized result in cases where some `Null` AID values were present. The algorithm has now been adapted to provide the largest possible safe aggregate count.