Special treatment of IsNull
We previously parsed and treated
`value IS NOT NULL` as being the same
as `not (value = null)`. Comparisons with
null do not however yield the expected result,
and as such the system didn't have the
desired effect.

This commit introduces IsNull as an Expression.
Basically the equivalent of the `IS NULL` syntax
in SQL.