Move the parser to a separate project
The parser is the only module depending on fparsec. FParsec doesn't compile to JS using Fable because of it's C# dependencies... this was the simplest way I could find to quickly get our system compiling to JavaScript in order for us to make a test anonymizer desktop application based on reference.