Update the noise generation mechanism to latest Elm spec.
For suppression and aggregation noise, we now use 2 noise layers,
one for the bucket and one for the AIDs.
For flattening, we use a seed generated from the max group of AIDs.
For each seed, we generate a single random value without using a PRNG.
A mechanism to create separate seeds for different steps is added.